refactor: use 64bit ID instead of 32
add get current user routes

+++ internal/api/dto/auth.dto.go
@@ -23,6 +23,6 @@ type UserAuthClaims struct {
 }
 
 type UserAuthPayload struct {
-	UserID          uint `json:"userID,omitempty"`
-	IsEmailVerified bool `json:"isEmailVerified"`
+	UserID          uint64 `json:"userID,omitempty"`
+	IsEmailVerified bool   `json:"isEmailVerified"`
 }
