+++ internal/models/token.model.go
@@ -7,7 +7,7 @@ import (
 
 type Token struct {
 	Token     string           `gorm:"primaryKey"`
-	UserID    uint             `gorm:"not null;index"`
+	UserID    uint64           `gorm:"not null;index"`
 	User      User             `gorm:"foreignKey:UserID;constraint:OnDelete:CASCADE;"`
 	Type      consts.TokenType `gorm:"not null"`
 	ExpiresAt time.Time        `gorm:"not null"`
